refactor(Application, CalculatorController): view와 model을 제어할 Controller 패키지 생성
view와 model간에 통신을 제어하는 controller를 생성하는 리팩토링을 진행했습니다
